Add files via upload
Does everything by running "python incScrape.py". Since this was made late at night and this is the first version, the user must change the ticker name in the code. I found the command line args to be annoying in my current testing environment. Should print Pandas dataframe with income data and uses matplotlib to plot the Gross Profit from the previous 4 years.